refactor(run-protocol): extract vaultKit module
